requirements.txt
numpy
matplotlib
PyMuPDF
requests
Pillow
mplcursors
opencv-python
io
BytesIO